Use new route for getting all other members; add refresh to forms and histories
