+++ src/routes/groups/index.tsx
@@ -1,3 +1,4 @@
+import { Settlement } from "@/routes/settlement";
 import GroupPage from "./group-id";
 import NewGroup from "./new-group";
 import { Purchase } from "@/routes/purchase";
@@ -16,6 +17,10 @@ const groupsRoutes = {
     {
       path: ":id/purchase",
       element: <Purchase />,
+    },
+    {
+      path: ":id/settlement",
+      element: <Settlement />,
     },
   ],
 };
